Improve Parser Performance by ~75%
Profiling the code showed that ~75% of the time was spent in the BufferedByteReader.pop() method. Almost all of that time was spent removing the read character from the buffer.

By storing an index, and removing elements from the buffer only when new elements are read, this bottleneck was removed.

Additionally, the peek(at:) method was rewritten to allow arbitraryly large lookaheads.